refactor: merge stacked list tests
