+++ arithmetization/src/test/java/net/consensys/linea/zktracer/containers/StackedListTests.java
@@ -17,14 +17,33 @@ package net.consensys.linea.zktracer.containers;
 
 import static org.assertj.core.api.Assertions.assertThat;
 
+import java.math.BigInteger;
+
+import com.google.common.collect.ImmutableList;
 import lombok.RequiredArgsConstructor;
 import net.consensys.linea.zktracer.container.ModuleOperation;
 import net.consensys.linea.zktracer.container.stacked.list.StackedList;
+import net.consensys.linea.zktracer.module.add.AddOperation;
+import net.consensys.linea.zktracer.opcode.OpCode;
+import org.apache.tuweni.bytes.Bytes;
+import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
 
 public class StackedListTests {
+  private static final AddOperation ONE_PLUS_ONE =
+      new AddOperation(
+          OpCode.ADD,
+          Bytes.wrap(BigInteger.ONE.toByteArray()),
+          Bytes.wrap(BigInteger.ONE.toByteArray()));
+
+  private static final AddOperation ONE_PLUS_TWO =
+      new AddOperation(
+          OpCode.ADD,
+          Bytes.wrap(BigInteger.ONE.toByteArray()),
+          Bytes.wrap(BigInteger.TWO.toByteArray()));
+
   @RequiredArgsConstructor
-  private class IntegerModuleOperation extends ModuleOperation {
+  private static class IntegerModuleOperation extends ModuleOperation {
     private final int x;
 
     @Override
@@ -47,5 +66,35 @@ public class StackedListTests {
 
     state.pop();
     assertThat(state.lineCount()).isEqualTo(1);
+  }
+
+  @Test
+  public void push() {
+    StackedList<AddOperation> chunks = new StackedList<>();
+    chunks.enter();
+
+    chunks.add(ONE_PLUS_ONE);
+    chunks.add(ONE_PLUS_ONE);
+    chunks.add(ONE_PLUS_ONE);
+    Assertions.assertEquals(3, chunks.size());
+    chunks.pop();
+    Assertions.assertEquals(0, chunks.size());
+  }
+
+  @Test
+  public void multiplePushPop() {
+    StackedList<AddOperation> chunks = new StackedList<>();
+    chunks.enter();
+    chunks.add(ONE_PLUS_ONE);
+    chunks.add(ONE_PLUS_ONE);
+    Assertions.assertEquals(2, chunks.size());
+
+    chunks.enter();
+    chunks.add(ONE_PLUS_ONE);
+    Assertions.assertEquals(3, chunks.size());
+    chunks.add(ONE_PLUS_TWO);
+    Assertions.assertEquals(4, chunks.size());
+    chunks.pop();
+    Assertions.assertEquals(2, ImmutableList.copyOf(chunks.iterator()).size());
   }
 }
